added error handling to activities

+++ backend/src/feature_user_activities/activities.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { BadRequestException, Injectable } from "@nestjs/common";
 import { ActivitiesRepository, ActivitiesDto } from "./activities.repository.js";
 
 @Injectable()
@@ -6,6 +6,9 @@ export class ActivitiesService {
     constructor(private readonly activitiesRepository: ActivitiesRepository) {}
 
     async remove(id: number) {
+
+        if (!id) throw new BadRequestException("Couldn't find the activity")
+
         return this.activitiesRepository.remove(id);
     }
 
@@ -24,10 +27,10 @@ export class ActivitiesService {
 
     async update(id: number, activitiesDto: ActivitiesDto) {
 
-        if (!id) throw new Error("Couldn't find the activity");
+        if (!id) throw new BadRequestException("Couldn't find the activity")
 
-        if (!activitiesDto.week && !activitiesDto.title && !activitiesDto.description && !activitiesDto.week && !activitiesDto.end) {
-            throw new Error("You're missing some fields")
+        if (!activitiesDto.week || !activitiesDto.title || !activitiesDto.description || !activitiesDto.week || !activitiesDto.end) {
+            throw new BadRequestException("You are missing some fields")
         }
 
         return this.activitiesRepository.update(id, activitiesDto);
